1. удален файл functions.js 2. изменен файл index.html (script) 3. разбитие кода на 3 модуля: data.js, util.js и main.js
